Adds rewriter library
Adds library which can peform rewriting using an FST. That is, given a
file containing a list of strings (one per line) it applies the rule and
computes the shortest path, then stringifies the result.

The user specifies the path to the rewrite rule FST, the path to the list of
strings, and the token type (or a symbol table).